chore(NotionIntegration): add green test: should create a page in a table with a title property

+++ tests/__helpers__/teardown.ts
@@ -3,6 +3,8 @@ import { join } from 'path'
 import { container } from './setup'
 
 async function globalTeardown() {
+  const isUnitTest = process.argv.includes('tests/unit')
+  if (isUnitTest) return
   await container.stop()
   await fs.emptyDir(join(process.cwd(), 'tmp'))
 }
